restructure condition_preview helpers: relocate, dedup, drop dead code
Helpers had drifted into a board_utils grab-bag with cross-file
duplication and a dead plan-expansion subtree. Audit + cleanup:

Relocations (one import home per symbol, no re-exports):
- pathClear + squareCompatibleOrEmpty → geometry_utils
- placeKingsIfAbsent + anyKingIsAdjacentTo + squaresAreAdjacent →
  king_placement (tests moved with them)
- pathClearOnPieces → sliderPathClear (clearer name + slider-step rewrite)
- board_utils sheds its placePiece + respect_caps imports, removing a
  latent board_utils → respect_caps → relations → geometry_utils cycle

Placement-legality dedup (canonical in piece_placement):
- legalPlacementForSpecies, teamHasKing, pawnCount, PAWN_CAP_PER_TEAM
  defined once in piece_placement; 16 sources + 2 tests repointed there
- board_utils drops the duplicates; MAX_PAWNS_PER_TEAM / pawnsOnTeam
  names retired

Dead code:
- delete square() (board_utils)
- delete expandRelationalPlanSources + its helper subtree from plan.js
  and generation_plan.js, its test, the now-orphaned value_source_options.js,
  and the dead VALUE_COMPARISON_METRIC constant

satisfiesComparator now throws on an unknown positionComparator instead
of silently treating it as equal_to
