+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/relations/relation_helpers.js
@@ -1,8 +1,6 @@
-import {
-  ALL_POSITIONS, WEIGHTED_SPECIES_DISTRIBUTION, pieceCode, legalPlacementForSpecies
-} from 'editorV2/panels/condition_preview/shared/board_utils'
+import { ALL_POSITIONS, WEIGHTED_SPECIES_DISTRIBUTION, pieceCode } from 'editorV2/panels/condition_preview/shared/board_utils'
 import { materialValue } from 'gameplay/board_query_utils'
-import { placePiece } from 'editorV2/panels/condition_preview/shared/piece_placement'
+import { placePiece, legalPlacementForSpecies } from 'editorV2/panels/condition_preview/shared/piece_placement'
 import { respectsAllCaps } from 'editorV2/panels/condition_preview/forward_proposition/respect_caps'
 import { regionPossiblyContains } from '../region'
