+++ app/javascript/editorV2/panels/condition_preview/plans/generation_plan.js
@@ -9,7 +9,7 @@ import {
 import {
   COUNT_COMPARISON_METRIC, INDIVIDUAL_VALUE_METRIC, AGGREGATE_VALUE_METRIC, isValueMetric,
   EXACT_NUMBER_COMPARISON_SOURCE, PRIOR_BOARD_COMPARISON_SOURCE,
-  comparisonDescriptors, comparisonRequirements, comparisonRequirementsFromDescriptors
+  comparisonDescriptors, comparisonRequirements
 } from 'editorV2/panels/condition_preview/plans/comparison_requirements'
 
 const SUPPORTED_UNARY_ACTORS = new Set(['allied', 'enemy', 'moved_piece', 'enemy_moved_piece', 'captured_piece', 'enemy_captured_piece'])
@@ -59,14 +59,6 @@ const FILTER_LABELS = Object.freeze({
   captured_piece: 'Captured piece',
   enemy_captured_piece: 'Enemy captured piece'
 })
-
-const NON_KING_VALUE_SPECIES = Object.freeze(new Map([
-  [0, []],
-  [1, [Board.PAWN]],
-  [3, [Board.NIGHT, Board.BISHOP]],
-  [5, [Board.ROOK]],
-  [9, [Board.QUEEN]]
-]))
 
 function actorLabel(actor) {
   return FILTER_LABELS[actor] || actor
@@ -150,118 +142,6 @@ function buildRelationalPlan(payload, options = {}, teams = {}) {
     moveKinds: options.moveKinds || [MOVE_KIND_STANDARD, MOVE_KIND_CASTLE, MOVE_KIND_PROMOTION, MOVE_KIND_EN_PASSANT],
     relationParams: relationParams(payload)
   }
-}
-
-function valueSourceOptions(descriptor) {
-  switch (descriptor.source) {
-    case EXACT_NUMBER_COMPARISON_SOURCE:
-      return [{
-        resolvedTotal: Number(descriptor.total || 0),
-        constraints: {}
-      }]
-    case 'moved_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({
-          resolvedTotal: value,
-          constraints: { movedPieceSpeciesPool: speciesPool }
-        }))
-    case 'captured_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({
-        resolvedTotal: value,
-        constraints: { capturedPieceSpeciesPool: speciesPool }
-        }))
-    case 'enemy_moved_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({
-        resolvedTotal: value,
-        constraints: { enemyMovedPieceSpeciesPool: speciesPool }
-        }))
-    case 'enemy_captured_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({
-        resolvedTotal: value,
-        constraints: { enemyCapturedPieceSpeciesPool: speciesPool }
-        }))
-    default:
-      return []
-  }
-}
-
-function mergeConstraintValue(left, right) {
-  if (left === undefined) { return right }
-  if (right === undefined) { return left }
-  if (left.length === 0 || right.length === 0) {
-    return left.length === right.length ? [] : null
-  }
-  const intersection = left.filter(species => right.includes(species))
-  return intersection.length > 0 ? intersection : null
-}
-
-function mergeConstraints(base, extra) {
-  const merged = { ...base }
-  const keys = new Set([...Object.keys(base), ...Object.keys(extra)])
-
-  for (const key of keys) {
-    const value = mergeConstraintValue(base[key], extra[key])
-    if (value === null) { return null }
-    if (value !== undefined) { merged[key] = value }
-  }
-
-  return merged
-}
-
-function expandDescriptorVariants(descriptors) {
-  return descriptors.reduce((variants, descriptor) => {
-    const options = isValueMetric(descriptor.metric)
-      ? valueSourceOptions(descriptor)
-      : [{
-        resolvedTotal: descriptor.total,
-        constraints: {}
-      }]
-
-    const nextVariants = []
-    variants.forEach(variant => {
-      options.forEach(option => {
-        const constraints = mergeConstraints(variant.sourceConstraints, option.constraints)
-        if (!constraints) { return }
-        nextVariants.push({
-          descriptors: [
-            ...variant.descriptors,
-            {
-              ...descriptor,
-              resolvedTotal: option.resolvedTotal
-            }
-          ],
-          sourceConstraints: constraints
-        })
-      })
-    })
-    return nextVariants
-  }, [{
-    descriptors: [],
-    sourceConstraints: {}
-  }])
-}
-
-export function expandRelationalPlanSources(plan) {
-  if (!plan.comparisonDescriptors.some(descriptor => (
-    isValueMetric(descriptor.metric) &&
-    ![EXACT_NUMBER_COMPARISON_SOURCE, PRIOR_BOARD_COMPARISON_SOURCE].includes(descriptor.source)
-  ))) {
-    return [plan]
-  }
-
-  return expandDescriptorVariants(plan.comparisonDescriptors).map(variant => ({
-    ...plan,
-    comparisonDescriptors: variant.descriptors,
-    sourceConstraints: variant.sourceConstraints,
-    requirements: comparisonRequirementsFromDescriptors(variant.descriptors)
-  }))
 }
 
 function singularActorValuePool(pool, comparator, total) {
