+++ app/javascript/editorV2/panels/condition_preview/plans/plan.js
@@ -4,109 +4,9 @@ import {
   MOVE_KIND_STANDARD, MOVE_KIND_CASTLE, MOVE_KIND_PROMOTION, MOVE_KIND_EN_PASSANT
 } from 'editorV2/panels/condition_preview/shared/example_utils'
 import { SINGULAR_ACTORS } from 'bot_execution/actors'
-import {
-  isValueMetric,
-  EXACT_NUMBER_COMPARISON_SOURCE,
-  PRIOR_BOARD_COMPARISON_SOURCE,
-  comparisonRequirementsFromDescriptors,
-  valueComparisonAllowsEmpty
-} from 'editorV2/panels/condition_preview/plans/comparison_requirements'
-import { valueSourceOptions as kingInclusiveValueSourceOptions } from 'editorV2/panels/condition_preview/plans/value_source_options'
+import { valueComparisonAllowsEmpty } from 'editorV2/panels/condition_preview/plans/comparison_requirements'
 
 const ALL_MOVE_KINDS = Object.freeze([MOVE_KIND_STANDARD, MOVE_KIND_CASTLE, MOVE_KIND_PROMOTION, MOVE_KIND_EN_PASSANT])
-
-// ===== expandRelationalPlanSources =====
-
-const NON_KING_VALUE_SPECIES = Object.freeze(new Map([
-  [0, []],
-  [1, [Board.PAWN]],
-  [3, [Board.NIGHT, Board.BISHOP]],
-  [5, [Board.ROOK]],
-  [9, [Board.QUEEN]]
-]))
-
-function valueSourceOptions(descriptor) {
-  switch (descriptor.source) {
-    case EXACT_NUMBER_COMPARISON_SOURCE:
-      return [{ resolvedTotal: Number(descriptor.total || 0), constraints: {} }]
-    case 'moved_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({ resolvedTotal: value, constraints: { movedPieceSpeciesPool: speciesPool } }))
-    case 'captured_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({ resolvedTotal: value, constraints: { capturedPieceSpeciesPool: speciesPool } }))
-    case 'enemy_moved_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({ resolvedTotal: value, constraints: { enemyMovedPieceSpeciesPool: speciesPool } }))
-    case 'enemy_captured_piece':
-      return Array.from(NON_KING_VALUE_SPECIES.entries())
-        .filter(([value]) => value > 0)
-        .map(([value, speciesPool]) => ({ resolvedTotal: value, constraints: { enemyCapturedPieceSpeciesPool: speciesPool } }))
-    default:
-      return []
-  }
-}
-
-function mergeConstraintValue(left, right) {
-  if (left === undefined) { return right }
-  if (right === undefined) { return left }
-  if (left.length === 0 || right.length === 0) {
-    return left.length === right.length ? [] : null
-  }
-  const intersection = left.filter(species => right.includes(species))
-  return intersection.length > 0 ? intersection : null
-}
-
-function mergeConstraints(base, extra) {
-  const merged = { ...base }
-  const keys = new Set([...Object.keys(base), ...Object.keys(extra)])
-  for (const key of keys) {
-    const value = mergeConstraintValue(base[key], extra[key])
-    if (value === null) { return null }
-    if (value !== undefined) { merged[key] = value }
-  }
-  return merged
-}
-
-function expandDescriptorVariants(descriptors) {
-  return descriptors.reduce((variants, descriptor) => {
-    const options = isValueMetric(descriptor.metric)
-      ? kingInclusiveValueSourceOptions(descriptor)
-      : [{ resolvedTotal: descriptor.total, constraints: {} }]
-
-    const nextVariants = []
-    variants.forEach(variant => {
-      options.forEach(option => {
-        const constraints = mergeConstraints(variant.sourceConstraints, option.constraints)
-        if (!constraints) { return }
-        nextVariants.push({
-          descriptors: [...variant.descriptors, { ...descriptor, resolvedTotal: option.resolvedTotal }],
-          sourceConstraints: constraints
-        })
-      })
-    })
-    return nextVariants
-  }, [{ descriptors: [], sourceConstraints: {} }])
-}
-
-export function expandRelationalPlanSources(plan) {
-  if (!plan.comparisonDescriptors.some(descriptor => (
-    isValueMetric(descriptor.metric) &&
-    ![EXACT_NUMBER_COMPARISON_SOURCE, PRIOR_BOARD_COMPARISON_SOURCE].includes(descriptor.source)
-  ))) {
-    return [plan]
-  }
-
-  return expandDescriptorVariants(plan.comparisonDescriptors).map(variant => ({
-    ...plan,
-    comparisonDescriptors: variant.descriptors,
-    sourceConstraints: variant.sourceConstraints,
-    requirements: comparisonRequirementsFromDescriptors(variant.descriptors)
-  }))
-}
 
 // ===== Combined plan helpers =====
