+++ app/javascript/editorV2/panels/condition_preview/shared/__tests__/king_placement.test.js
@@ -2,11 +2,9 @@ import { beforeAll, beforeEach, describe, expect, it } from 'vitest'
 import Board from 'gameplay/board'
 import { controllingPositions } from 'gameplay/board_query_utils'
 
-import { placeKingDeliberately, placeKingInCheck, placeKingInStalemate } from 'editorV2/panels/condition_preview/shared/king_placement'
-import {
-  pieceCode, teamHasKing,
-  buildBoardFromLayout, buildLayoutFromPieces
-} from 'editorV2/panels/condition_preview/shared/board_utils'
+import { placeKingDeliberately, placeKingInCheck, placeKingInStalemate, placeKingsIfAbsent } from 'editorV2/panels/condition_preview/shared/king_placement'
+import { pieceCode, buildBoardFromLayout, buildLayoutFromPieces } from 'editorV2/panels/condition_preview/shared/board_utils'
+import { teamHasKing } from 'editorV2/panels/condition_preview/shared/piece_placement'
 
 const A1 = 0
 const D4 = 27
@@ -152,5 +150,32 @@ describe('placeKingInStalemate — empty board', () => {
 
   it('successfully constructs at least one stalemate across 30 seeds', () => {
     expect(successes).toBeGreaterThan(0)
+  })
+})
+
+describe('placeKingsIfAbsent — cap-respecting', () => {
+  it('returns null when every legal king placement would violate a count_range.max cap on white kings', () => {
+    const ctx = {
+      singulars: {},
+      propositions: [{
+        team: Board.WHITE,
+        frame: 'current',
+        species_set: new Set([Board.KING]),
+        region: { kind: 'all' },
+        count_range: { min: 0, max: 0 },
+        aggregate_value_range: { min: 0, max: Infinity },
+        aggregate_mobility_range: { min: 0, max: Infinity }
+      }]
+    }
+
+    expect(placeKingsIfAbsent(new Map(), () => 0, ctx)).toBeNull()
+  })
+
+  it('places both kings when no caps are present (default ctx)', () => {
+    const result = placeKingsIfAbsent(new Map(), () => 0)
+
+    expect(result).not.toBeNull()
+    expect(teamHasKing(result, Board.WHITE)).toBe(true)
+    expect(teamHasKing(result, Board.BLACK)).toBe(true)
   })
 })
