+++ app/javascript/editorV2/panels/condition_preview/shared/board_utils.js
@@ -1,14 +1,7 @@
 import Board from 'gameplay/board'
-import { placePiece } from './piece_placement'
-import { respectsAllCaps } from 'editorV2/panels/condition_preview/forward_proposition/respect_caps'
-
-export const MAX_PAWNS_PER_TEAM = 8
+import { legalPlacementForSpecies, pawnCount, PAWN_CAP_PER_TEAM } from './piece_placement'
 
 export const HOME_RANK = Object.freeze({ [Board.WHITE]: 0, [Board.BLACK]: 7 })
-
-export function square(value) {
-  return Board.gridCalculatorReverse(value)
-}
 
 export function pieceCode(team, species) {
   return `${team}${species}`
@@ -20,12 +13,6 @@ export function pieceTeam(piece) {
 
 export function pieceSpecies(piece) {
   return piece ? Board.parseSpecies(piece) : null
-}
-
-export function legalPlacementForSpecies(position, species) {
-  if (species !== Board.PAWN) { return true }
-  const rank = Board.rankIndex(position)
-  return rank !== 0 && rank !== 7
 }
 
 export function unique(values) {
@@ -106,20 +93,6 @@ export function orderedBlockerSpeciesFor(position, random) {
   return weightedShuffleSpecies(new Set(blockerSpeciesFor(position)), random)
 }
 
-export function pathClear(squares, fromIdx, untilIdx, pieces) {
-  for (let i = fromIdx; i < untilIdx; i += 1) {
-    if (pieces.has(squares[i])) { return false }
-  }
-  return true
-}
-
-export function squareCompatibleOrEmpty(pieces, pos, team, speciesSet) {
-  const existing = pieces.get(pos)
-  if (!existing) { return true }
-  if (existing.charAt(0) !== team) { return false }
-  return speciesSet.has(existing.slice(1))
-}
-
 export function buildLayoutFromPieces(pieces) {
   const layout = Array(64).fill(Board.EMPTY_SQUARE)
   pieces.forEach((piece, position) => {
@@ -146,25 +119,8 @@ export function layoutsMatch(left, right) {
   return true
 }
 
-export function teamHasKing(pieces, team) {
-  const kingCode = `${team}${Board.KING}`
-  for (const piece of pieces.values()) {
-    if (piece === kingCode) { return true }
-  }
-  return false
-}
-
-export function pawnsOnTeam(pieces, team) {
-  const pawnCode = `${team}${Board.PAWN}`
-  let count = 0
-  for (const piece of pieces.values()) {
-    if (piece === pawnCode) { count += 1 }
-  }
-  return count
-}
-
 export function legalEnrichmentSpecies(pieces, team) {
-  const excludePawns = pawnsOnTeam(pieces, team) >= MAX_PAWNS_PER_TEAM
+  const excludePawns = pawnCount(pieces, team) >= PAWN_CAP_PER_TEAM
   return WEIGHTED_SPECIES_DISTRIBUTION.filter(species => {
     if (species === Board.KING) { return false }
     if (species === Board.PAWN && excludePawns) { return false }
@@ -172,49 +128,3 @@ export function legalEnrichmentSpecies(pieces, team) {
   })
 }
 
-function squaresAreAdjacent(a, b) {
-  return (
-    Math.abs(Board.rankIndex(a) - Board.rankIndex(b)) <= 1 &&
-    Math.abs(Board.fileIndex(a) - Board.fileIndex(b)) <= 1 &&
-    a !== b
-  )
-}
-
-export function anyKingIsAdjacentTo(pieces, position) {
-  for (const [sq, piece] of pieces.entries()) {
-    if (
-      (piece === Board.WHITE_KING || piece === Board.BLACK_KING) &&
-      squaresAreAdjacent(sq, position)
-    ) {
-      return true
-    }
-  }
-  return false
-}
-
-export function placeKingsIfAbsent(pieces, random, ctx = { propositions: [] }) {
-  let result = pieces
-
-  for (const team of [Board.WHITE, Board.BLACK]) {
-    if (teamHasKing(result, team)) { continue }
-
-    const candidates = shuffled(
-      ALL_POSITIONS.filter(pos => !result.has(pos)),
-      random
-    )
-
-    let placed = false
-    for (const pos of candidates) {
-      if (anyKingIsAdjacentTo(result, pos)) { continue }
-      if (!respectsAllCaps(team, Board.KING, pos, ctx, result)) { continue }
-      const next = placePiece(result, pos, `${team}${Board.KING}`)
-      if (next === null) { continue }
-      result = next
-      placed = true
-      break
-    }
-    if (!placed) { return null }
-  }
-
-  return result
-}
