+++ bin/render-build.sh
@@ -6,4 +6,4 @@ bundle install
 bundle exec rails assets:precompile
 bundle exec rails assets:clean
 bundle exec rails db:migrate
-# bundle exec rails db:seed
+bundle exec rails db:seed
